Fixed bitmap image delete crash

+++ tgb.lua
@@ -1059,7 +1059,9 @@ local function addBitmap( oSurface, nPosx, nPosy, sBitmapPath, nOpacity )
   -- Image implementation
   function image.delete()
     for i=1, #pixels do
-      pixels[i].delete()
+      if pixels[i].delete then
+        pixels[i].delete()
+      end
     end
   end
